requirements.txt
SpeechRecognition
requests
pyttsx3
psycopg2.binary
wikipedia
autopep8
selenium
pylint
instaloader
pyaudio
text_to_speech